+++ packages/floop/example/example.dart
@@ -55,7 +55,56 @@ class ImageDisplay extends StatelessWidget with Floop {
         child: Icon(Icons.refresh),
         onPressed: () async {
           await fetchImage();
-          // Restart context transitions after image has been loaded.
+          // Restarting context transitions after the new image has loaded
+          // causes the new image to also transition from top to center.
+          Transitions.restart(context: context);
+        },
+      ),
+    );
+  }
+}
+
+class DynamicValues {
+  static Widget get image => floop['image'];
+  static set image(Widget widget) => floop['image'] = widget;
+}
+
+fetchImage2([String url = 'https://picsum.photos/300/200']) async {
+  if (_fetching) {
+    return;
+  }
+  _fetching = true;
+  DynamicValues.image = null; // Set to null while awaiting the response
+  final response = await http.get(url);
+
+  // The image is stored only when this is the last call to fetchImage,
+  DynamicValues.image = TransitionImage(Image.memory(response.bodyBytes));
+  _fetching = false;
+}
+
+class ImageDisplay2 extends StatelessWidget with Floop {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      // `floop['image']` is null while fetching an image. When the
+      // imaged is downloaded, an image widget is stored on `floop['image']`
+      // and the widget automatically updates.
+      body: DynamicValues.image == null
+          ? Center(
+              child: Text(
+                'Loading...',
+                textScaleFactor: 2,
+              ),
+            )
+          : Align(
+              alignment: Alignment(0, transition(2000, delayMillis: 800) - 1),
+              child: DynamicValues.image),
+      floatingActionButton: FloatingActionButton(
+        child: Icon(Icons.refresh),
+        onPressed: () async {
+          await fetchImage();
+          // Restarting context transitions after the new image has loaded
+          // causes the new image to also transition from top to center.
           Transitions.restart(context: context);
         },
       ),