+++ packages/floop/CHANGELOG.md
@@ -1,5 +1,7 @@
-## [0.4.0] - 2/09/2019
+## [0.1.0] - 2/09/2019
 
+- Adding `with Floop` is all that is required to enable Floop on a widget
+- No more [buildWithFloop], the build method is just [build]
 - Major refactor to core functionality. Listening is faster.
 - Eliminated [FloopLight], examples updated accordingly.
 - Updated the clicker example. Now it fetches an image.